pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "wheel>=0.40.0",
    "pybind11>=2.10.0",
    "cython>=0.29.21",
    "numpy>=1.19.5,<2.0",
    "setuptools-scm[toml]>=7.0.0",
    "docutils>=0.17",
    "tqdm>=4.62.3,<5.0.0",
    "missingno >=0.5.2,<0.6.0"
]
build-backend = "setuptools.build_meta"  # Use setuptools as the build backend

# dependencies 是运行时依赖，也就是用户 pip install sequenzo 时会自动装的内容
[project]
name = "sequenzo"
description = "A fast, scalable and intuitive Python package for social sequence analysis."
readme = "README.md"
authors = [
    { name = "Yuqi Liang", email = "yuqi.liang.1900@gmail.com" },
    { name = "Xinyi Li", email = "1836724126@qq.com" },
    { name = "Jan Heinrich Ernst Meyerhoff-Liang", email = "jan.meyerhoff1@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.9,<3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
version = "0.1.11"
dependencies = [
    "numpy>=1.19.5,<2.0",
    "pandas>=1.2.5",
    "matplotlib>=3.4.3",
    "seaborn>=0.11.2",
    "Pillow>=8.3.2",
    "pybind11>=2.6.0",
    "cython>=0.29.21",
    "scipy>=1.6.3",
    "scikit-learn>=0.24.2",
    "fastcluster>=1.2.6",
    "joblib>=1.0.1",
    "docutils>=0.17",
    "tqdm>=4.62.3,<5.0.0",
    "missingno>=0.5.2,<0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "flake8>=3.9.2",
    "memory-profiler>=0.58.0"
]

[project.urls]
Homepage = "https://github.com/Liang-Team/Sequenzo"
Documentation = "https://sequenzo.yuqi-liang.tech"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"sequenzo.datasets" = ["*.csv"]
"sequenzo.dissimilarity_measures" = [
  "src/xsimd/include/**/*.hpp",
  "src/xsimd/include/**/*.h"
]


[tool.setuptools.packages.find]
# 这告诉 setuptools：sequenzo 包目录就直接在当前目录下（不是 sequenzo/sequenzo 里）
where = ["."]
include = ["sequenzo*"]

[tool.pybind11]
include = ["pybind11"]